Integer statement.

+++ Builder/Statements/const_int_statement.h
@@ -3,7 +3,7 @@
 class ConstIntStatement : public Statement
 {
 public:
-    ConstIntStatement(string line, int value);
+    ConstIntStatement(string line, long long value);
     static ConstIntStatement *Build(string line);
 
 private:
